app/training/requirements.txt
numpy==1.25.2
tqdm==4.66.1
torch>=1.6.0
sentence-transformers==2.2.2
wandb==0.15.8
pandas==2.0.3
scikit-learn==1.3.0
transformers==4.31.0
faiss-cpu
